+++ src/features/auth/hooks/use-fetch-user-by-google-token.ts
@@ -1,16 +1,13 @@
-import { verifyGoogleToken } from "@/src/api/authentification/verify-google-token";
-import { User } from "@/types/user";
+import {
+  verifyGoogleToken,
+  VerifyGoogleTokenResponse,
+} from "@/src/api/authentification/verify-google-token";
 import { MutationOptions, useMutation } from "@tanstack/react-query";
 
-export type AuthResponse = {
-  token: string;
-  user: User;
-};
-
 export const useFetchUserByGoogleToken = (
-  props: MutationOptions<AuthResponse, Error, string>
+  props: MutationOptions<VerifyGoogleTokenResponse, Error, string>
 ) => {
-  return useMutation<AuthResponse, Error, string>({
+  return useMutation<VerifyGoogleTokenResponse, Error, string>({
     mutationFn: (googleToken: string) => verifyGoogleToken(googleToken),
     ...props,
   });